Moved tests and added case for Latest block number
