+++ crml/eth-bridge/src/tests.rs
@@ -15,14 +15,9 @@
 
 use super::BridgePaused;
 use crate::mock::*;
-use crate::types::{EthBlock, EthHash, LatestOrNumber, TransactionReceipt};
-use crate::{
-	types::{BridgeEthereumRpcApi, EventProofId},
-	Error, Module,
-};
+use crate::{types::EventProofId, Error, Module};
 use cennznet_primitives::eth::crypto::AuthorityId;
-use crml_support::{EthAbiCodec, EventClaimVerifier, H160, U256};
-use ethereum_types::U64;
+use crml_support::{EthAbiCodec, EventClaimVerifier, H160};
 use frame_support::{
 	assert_noop, assert_ok,
 	dispatch::DispatchError,
@@ -137,18 +132,6 @@ fn eth_client_http_request() {
 			..Default::default()
 		});
 	}
-
-	// Test
-	// t.execute_with(|| {
-	// 	let response =
-	// 		<MockEthereumRpcClient as BridgeEthereumRpcApi>::query_eth_client(request_body).expect("got response");
-	// 	assert_eq!(
-	// 		serde_json::from_slice::<'_, TestRequest>(response.as_slice()).unwrap(),
-	// 		TestRequest {
-	// 			message: "hello cennznet".to_string()
-	// 		}
-	// 	);
-	// })
 }
 
 #[test]
@@ -346,77 +329,6 @@ fn set_delayed_event_proofs_per_block_not_root_should_fail() {
 			DispatchError::BadOrigin
 		);
 		assert_eq!(Module::<TestRuntime>::delayed_event_proofs_per_block(), 5);
-	});
-}
-
-#[test]
-fn get_block_by_number_mock_works() {
-	ExtBuilder::default().build().execute_with(|| {
-		let block_number: u32 = 120;
-		let block_hash: H256 = H256::from_low_u64_be(121);
-		let timestamp: U256 = U256::from(122);
-
-		let mock_block = EthBlock {
-			number: Some(U64::from(block_number)),
-			hash: Some(block_hash),
-			timestamp,
-			parent_hash: Default::default(),
-			nonce: None,
-			sha3_uncles: Default::default(),
-			logs_bloom: None,
-			transactions_root: Default::default(),
-			state_root: Default::default(),
-			receipts_root: Default::default(),
-			miner: Default::default(),
-			difficulty: Default::default(),
-			total_difficulty: Default::default(),
-			extra_data: vec![],
-			size: Default::default(),
-			gas_limit: Default::default(),
-			gas_used: Default::default(),
-			transactions: vec![],
-			uncles: vec![],
-		};
-
-		MockEthereumRpcClient::mock_block_response_at(block_number, mock_block.clone());
-
-		let result =
-			<MockEthereumRpcClient as BridgeEthereumRpcApi>::get_block_by_number(LatestOrNumber::Number(block_number))
-				.unwrap();
-		assert_eq!(Some(mock_block), result);
-	});
-}
-
-#[test]
-fn get_transaction_receipt_mock_works() {
-	ExtBuilder::default().build().execute_with(|| {
-		let block_number: u32 = 120;
-		let block_hash: H256 = H256::from_low_u64_be(121);
-		let tx_hash: EthHash = H256::from_low_u64_be(122);
-
-		let mock_tx_receipt = TransactionReceipt {
-			block_hash,
-			block_number: U64::from(block_number),
-			contract_address: None,
-			cumulative_gas_used: Default::default(),
-			effective_gas_price: None,
-			from: Default::default(),
-			gas_used: None,
-			logs: vec![],
-			status: None,
-			to: None,
-			transaction_hash: tx_hash,
-			transaction_index: Default::default(),
-			root: None,
-			logs_bloom: Default::default(),
-			transaction_type: None,
-			removed: false,
-		};
-
-		MockEthereumRpcClient::mock_transaction_receipt_for(tx_hash, mock_tx_receipt.clone());
-
-		let result = <MockEthereumRpcClient as BridgeEthereumRpcApi>::get_transaction_receipt(tx_hash).unwrap();
-		assert_eq!(Some(mock_tx_receipt), result);
 	});
 }
